chore: add StackBlitz fallback
