+++ docs/components/StackBlitz/StackBlitz.module.css
@@ -1,0 +1,8 @@
+.root {
+  display: block;
+  margin-top: 1rem;
+}
+
+.root > * + * {
+  margin-top: 1rem;
+}
