+++ docs/components/StackBlitz/index.ts
@@ -1,0 +1,1 @@
+export * from "./StackBlitz";
